+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Unreleased
 
 - OpenAI: Add `cyber_policy` to "content_filter" stop reason
-- Timelines: New `BranchEvent` to delineate timeline branches.
+- Timelines: `BranchEvent` and `timeline_branch()` to delineate timeline branches.
+- Timelines: Consolidate `TimelineBranch` into `TimelineSpan` via `forked_at` property.
 
 ## 0.3.202 (31 March 2026)